refactor(harvest): Move AccountForm and AccountFormError 🛠
